suggestions for changes to createsurvey help file
changed the recipients description to a list of lists.
changed the example function calls so that recipients are lists of
lists.
Changed the description of the email_body_to indicate the three required
links (unless user has a plan that allows ignoring the [FooterLink]
Corrected the reference to deprecated [RemoveLink], replaced with
current [OptOutLink]